Create Question 1
While purchasing certain items, a discount of 10% is offered if the quantity purchased is more than 50. If the quantity purchased is between 25 and 50 (inclusive), a discount of 5% is offered. If the quantity purchased is less than 25, no discount is offered. Write a program to calculate the total expenses given the quantity and price per item entered through the keyboard.